spacewalk-java: Adjust business-logic of Cobbler-API consumers
This commits enables the spacewalk-java code to work with the Cobbler 3.3.3
version of the Client that uses Optionals in properties that have resolved
values.

+++ java/code/src/com/redhat/rhn/common/conf/ConfigDefaults.java
@@ -540,8 +540,8 @@ public class ConfigDefaults {
      * Returns the default virt disk size in GBs
      * @return the virt disk size
      */
-    public int getDefaultVirtDiskSize() {
-        return Config.get().getInt(VIRT_DISK, 3);
+    public double getDefaultVirtDiskSize() {
+        return Config.get().getDouble(VIRT_DISK, 3.0);
     }
 
     /**
